Revert the butotn style; Move the trader history to futures; Rename the pagination style
