+++ packages/app/src/components/Table/Pagination.tsx
@@ -38,11 +38,10 @@ const Pagination: FC<PaginationProps> = React.memo(
 
 		const firstPage = () => setPage(0)
 		const toLastPage = () => setPage(pageCount - 1)
-		// eslint-disable-next-line no-console
-		console.log(`extra`, !!extra)
+
 		return (
 			<>
-				<PaginationContainer compact={compact} $extra={!!extra}>
+				<PaginationContainer $compact={compact} $bottomBorder={!!extra}>
 					<ArrowButtonContainer>
 						<ArrowButton onClick={firstPage} disabled={!canPreviousPage}>
 							<LeftEndArrowIcon />
@@ -85,14 +84,18 @@ const PageInfo = styled.span`
 	color: ${(props) => props.theme.colors.selectedTheme.gray};
 `
 
-const PaginationContainer = styled(GridDivCenteredCol)<{ compact: boolean; $extra: boolean }>`
+const PaginationContainer = styled(GridDivCenteredCol)<{
+	$compact: boolean
+	$bottomBorder: boolean
+}>`
 	grid-template-columns: auto 1fr auto;
-	padding: ${(props) => (props.compact ? '10px' : '15px')} 12px;
+	padding: ${(props) => (props.$compact ? '10px' : '15px')} 12px;
 	border-bottom-left-radius: 4px;
 	border-bottom-right-radius: 4px;
 	justify-items: center;
 	border-top: ${(props) => props.theme.colors.selectedTheme.border};
-	border-bottom: ${(props) => (props.$extra ? props.theme.colors.selectedTheme.border : 'none')};
+	border-bottom: ${(props) =>
+		props.$bottomBorder ? props.theme.colors.selectedTheme.border : 'none'};
 
 	${media.lessThan('lg')`
 		border-radius: 0px;
